+++ CHANGES.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented here. The format is based
 on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `fallbackComponent` (node) property can be used to render alternative component
+  when the requested `component` does not exist
+- log error when `<CatalogComponent />` is used without a `<CatalogProvider />`
+  context
+
 ## 2019/02/02 0.3.0
 
 ### Changed